requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
numpy
pandas
firebase-admin